+++ src/Nodes.js
@@ -2,33 +2,8 @@ import React, { PureComponent } from 'react';
 import { graphql } from 'react-apollo';
 import PropTypes from 'prop-types';
 import queryNodes from './queries';
-
-const NO_FOUND = 'No found';
-const LABEL_NAME = 'Label is ';
-
-const findLabel = (node, id) => {
-  if (node && node.id === id) {
-    return node.label;
-  }
-
-  let label = NO_FOUND;
-  const { children } = node;
-
-  if (children) {
-    const { length } = children;
-    let i = 0;
-
-    for (i; i < length; i += 1) {
-      label = findLabel(children[i], id);
-
-      if (label !== NO_FOUND) {
-        return label;
-      }
-    }
-  }
-
-  return label;
-};
+import findLabel from './findLabel';
+import { NO_FOUND, LABEL_NAME } from './constants';
 
 class Nodes extends PureComponent {
   constructor(props) {
