+++ src/constants.js
@@ -1,0 +1,2 @@
+export const NO_FOUND = 'No found';
+export const LABEL_NAME = 'Label is ';
